+++ tests/test_pydsmcli.py
@@ -1,0 +1,6 @@
+import pydsm
+
+def test_pydsm_cli():
+    ''' test the pydsm command line interface works correctly'''
+    pass
+